ingest-lamdba/requirements.txt
numpy==1.18.4
pandas==1.0.4
requests==2.23.0
boto3==1.13.19
fil_finder==1.7
Pillow==7.1.2
scikit-image==0.17.2
# scipy==1.4.1
# gdal==3.1.0
